backend/requirements.txt
flask==2.3.2
flask-cors==3.0.10
opencv-python-headless==4.7.0.72
rembg==2.0.38
gunicorn==20.1.0
python-dotenv==1.0.0
dnspython==2.3.0
pymongo==4.3.3
numpy==1.24.3
Pillow==9.5.0